nginx: handle port 80 for boxd.it and letterboxd.com
We need to handle boxd.it 80 so we can redirect to https before going through Varnish. Otherwise Varnish redirects to https as we tell Varnish it’s a request for http://letterboxd.com/. And our Apache servers don’t all know about boxd.it… and soon none of them will!