Infinispan: enable state transfer but don’t wait for it
If we have state transfer on startup then we don’t lose the cluster data, so we don’t suffer such a massive hit due to loss of cache.

At the time I couldn’t use await-initial-transfer as it took a long time to complete (maybe because at the time our cluster was out of memory and GCing) and seemed to run into a 5 minute startup timeout.